requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
streamlit==1.34.0
plotly==5.22.0
requests==2.31.0
apscheduler==3.10.4
faker==24.11.0
joblib
loguru
pytest
pytest-cov
pytest-asyncio
locust
